Remove data request info on password reset
