+++ src/resolvers/Mutation.ts
@@ -1,11 +1,7 @@
 import bcrypt from "bcrypt";
-import { randomBytes } from "crypto";
 import { validate } from "isemail";
-import moment from "moment";
-import { promisify } from "util";
-import { prisma, Prisma, UserCreateInput } from "../generated/prisma-client";
-import { transporter } from "../util";
-import { createDataRequest } from "./DataRequest";
+import { Prisma, UserCreateInput } from "../generated/prisma-client";
+import { requestReset, resetPassword } from "./resetPassword";
 
 // Defining the context interface
 interface IContext {
@@ -15,105 +11,11 @@ interface IContext {
 
 // Defining  Mutation methods
 const Mutation = {
-  /**
-   * Create a data request.
-   * A user need to be logged in to create a dat request
-   * @args
-   *      parts:
-   *          An array of data request part
-   *          It can be associate with observation/ it only have a part of the observation
-   *      part.ids:
-   *          An array of Data files ids
-   *          This files id's need to exist within the prisma database.
-   *          i.e you can only request data files that exist anything else will fail
-   * @return
-   *      New data request that has just been created or an error
-   */
-  createDataRequest: (root: any, args: any, ctx: IContext) =>
-    createDataRequest(args, ctx),
+  requestPasswordReset: (root: any, { email }: any, ctx: any) =>
+    requestReset(email),
 
-  async requestPasswordReset(root: any, { email }: any, ctx: any) {
-    // Find a user with given email address
-    const user = await prisma.user({
-      email
-    });
-    if (!user) {
-      throw new Error(`No user with email ${email}`);
-    }
-
-    // Create a reset token
-    const randomBytesPromisified = promisify(randomBytes);
-    const passwordResetToken = (await randomBytesPromisified(30)).toString(
-      "hex"
-    );
-
-    // Create a reset token expiry
-    const passwordResetTokenExpiry = moment(Date.now())
-      .add(2, "hours")
-      .toDate();
-
-    // Update the user in prisma with token and expiry date
-    const updatedUser = await prisma.updateUser({
-      data: {
-        passwordResetToken,
-        passwordResetTokenExpiry
-      },
-      where: { email }
-    });
-
-    // in case token is not unique which is very rare or prisma fail to update user fro some reasons
-    if (!updatedUser) {
-      throw new Error(`Fail to generate a reset token, request again`);
-    }
-    try {
-      const url = `${
-        process.env.HOST
-      }/auth/reset-password/${passwordResetToken}`;
-      await transporter.sendMail({
-        html: `Please to reset your password click: <a href="${url}">${url}</a> <br/>`,
-        subject: "Reset password SAAO/SALT Data archive",
-        to: user.email
-      });
-    } catch (e) {
-      throw new Error(`Fail to send reset token to email: ${email}`);
-    }
-    return user;
-  },
-
-  async resetPassword(
-    root: any,
-    { token, password }: any,
-    { prisma }: IContext
-  ) {
-    if (!(password.length > 6)) {
-      throw new Error(`The password must be at least 7 characters long.`);
-    }
-
-    // get the user with the token
-    const user = await prisma.user({ passwordResetToken: token });
-    if (!user) {
-      throw new Error(`Fail to reset password of unknown token`);
-    }
-    if (
-      user.passwordResetTokenExpiry &&
-      moment(user.passwordResetTokenExpiry) <= moment(Date.now())
-    ) {
-      throw new Error("Token is expired");
-    }
-    const hashedPassword = await bcrypt.hash(password, 10);
-
-    // updating the user with the new password and expire the token
-    await prisma.updateUser({
-      data: {
-        password: hashedPassword,
-        passwordResetTokenExpiry: moment(Date.now()).toDate()
-      },
-      where: {
-        passwordResetToken: token
-      }
-    });
-    return user;
-  },
+  resetPassword: (root: any, { token, password }: any, ctx: IContext) =>
+    resetPassword(token, password),
 
   /**
    * Register a new user.
