+++ src/resolvers/index.ts
@@ -1,10 +1,8 @@
-import { DataRequestsTypes } from "./DataRequest";
 import { Mutation } from "./Mutation";
 import { Query } from "./Query";
 
 // Defining resolvers
 const resolvers = {
-  ...DataRequestsTypes,
   Mutation,
   Query
 };
